refactor: move url_safe/1 to Utils where it belongs
