+++ lib/fun_with_flags/ui/router.ex
@@ -141,7 +141,7 @@ defmodule FunWithFlags.UI.Router do
       FunWithFlags.disable(flag_name, for_actor: actor)
     end
 
-    redirect_to conn, "/flags/#{name}#actor_#{Templates.url_safe(actor_id)}"
+    redirect_to conn, "/flags/#{name}#actor_#{Utils.url_safe(actor_id)}"
   end
 
 
@@ -169,7 +169,7 @@ defmodule FunWithFlags.UI.Router do
       FunWithFlags.disable(flag_name, for_group: group_name)
     end
 
-    redirect_to conn, "/flags/#{name}#group_#{Templates.url_safe(group_name)}"
+    redirect_to conn, "/flags/#{name}#group_#{Utils.url_safe(group_name)}"
   end
 
 
@@ -209,7 +209,7 @@ defmodule FunWithFlags.UI.Router do
         else
           FunWithFlags.disable(flag_name, for_actor: actor)
         end
-        redirect_to conn, "/flags/#{name}#actor_#{Templates.url_safe(actor_id)}"
+        redirect_to conn, "/flags/#{name}#actor_#{Utils.url_safe(actor_id)}"
       {:fail, reason} ->
         {:ok, flag} = Utils.get_flag(name)
         body = Templates.details(conn: conn, flag: flag, actor_error_message: "The actor ID #{reason}.")
@@ -232,7 +232,7 @@ defmodule FunWithFlags.UI.Router do
         else
           FunWithFlags.disable(flag_name, for_group: group_name)
         end
-        redirect_to conn, "/flags/#{name}#group_#{Templates.url_safe(group_name)}"
+        redirect_to conn, "/flags/#{name}#group_#{Utils.url_safe(group_name)}"
       {:fail, reason} ->
         {:ok, flag} = Utils.get_flag(name)
         body = Templates.details(conn: conn, flag: flag, group_error_message: "The group name #{reason}.")
